examples/09-computer-use-browser/requirements.txt
google-adk>=1.31,<2
playwright>=1.52
